qcad: symlink system QT libs instead of copying them
The QCAD build system copies QT libs into its install which means they
will not be updated when the package that provides it gets updated
unless QCAD gets rebuilt as well.

Sadly, QCAD cannot directly use the QT plugins where they are because it
will find the QT designer plugins, which are apparently not compatible.
So, we modify the build setup so that it symlinks the QT libs instead
of copying them. Plus, we make sure to install of the directories that
the build system puts those (now symlinked) libs in.